+++ internal/model/OrderAccrualModel.go
@@ -1,7 +1,7 @@
 package model
 
 type OrdersAccrualModel struct {
-	OrderId *string `json:"order"`
-	Accrual *int64  `json:"accrual"`
-	Status  *string `json:"status"`
+	OrderId *string  `json:"order"`
+	Accrual *float64 `json:"accrual"`
+	Status  *string  `json:"status"`
 }
